feat : 프로젝트 기능 초기 세팅 및 패키징 정리
