+++ src/main/java/com/isfive/usearth/domain/project/ProjectController.java
@@ -1,0 +1,4 @@
+package com.isfive.usearth.domain.project;
+
+public class ProjectController {
+}
