+++ src/main/java/com/isfive/usearth/domain/project/ProjectRepository.java
@@ -1,0 +1,7 @@
+package com.isfive.usearth.domain.project;
+
+import com.isfive.usearth.domain.project.entity.Project;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProjectRepository extends JpaRepository<Project, Long> {
+}
