+++ src/main/java/com/isfive/usearth/domain/project/ProjectService.java
@@ -1,0 +1,12 @@
+package com.isfive.usearth.domain.project;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class ProjectService {
+
+
+
+}
