+++ src/main/java/com/isfive/usearth/domain/project/entity/Option.java
@@ -1,4 +1,4 @@
-package com.isfive.usearth.domain.funding.entity;
+package com.isfive.usearth.domain.project.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -9,17 +9,19 @@ import java.util.List;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class RewardSku {
+@Table(name = "options")
+public class Option {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Integer stock;
+    private String name; // 색상, 사이즈
 
     @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "reward_id")
     private Reward reward;
 
-    @OneToMany(mappedBy = "rewardSku")
-    private List<SkuValue> skuValues = new ArrayList<>();
+    @OneToMany(mappedBy = "option")
+    private List<OptionValue> optionValues = new ArrayList<>();
 }
