+++ src/main/java/com/isfive/usearth/domain/project/entity/Tag.java
@@ -1,4 +1,4 @@
-package com.isfive.usearth.domain.funding.entity;
+package com.isfive.usearth.domain.project.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -6,16 +6,15 @@ import lombok.NoArgsConstructor;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class OptionValue {
+public class Tag {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(name = "option_value")
-    private String value;
+    private String name;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "options_id")
-    private Option option;
+    @JoinColumn(name = "project_id")
+    private Project project;
 }
