+++ src/test/java/com/isfive/usearth/domain/project/ProjectControllerTest.java
@@ -1,0 +1,7 @@
+package com.isfive.usearth.domain.project;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class ProjectControllerTest {
+
+}
